Upgrade Next.js to 15.2.0-canary.3, add Loader component, and implement ProductDetails with carousel support
