+++ package.json
@@ -26,7 +26,7 @@
     "clsx": "^2.1.1",
     "embla-carousel-react": "^8.5.2",
     "lucide-react": "^0.469.0",
-    "next": "15.1.3",
+    "next": "^15.2.0-canary.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
